hotfix/18824/dao-hotfix added component commentFooter
