+++ components/app/Pages/Discussions/CommentField3.vue
@@ -57,42 +57,11 @@
           </div>
         </div>
       </div>
-      <validation-observer
-        v-slot="{handleSubmit, validated, passed, invalid}"
-      >
-        <!--  Bottom -->
-        <div
-          v-if="isReply"
-          class="comment__footer footer"
-        >
-          <base-btn
-            class="footer__btn hide"
-            :disabled="!validated || !passed || invalid"
-          >
-            <template v-slot:left>
-              <span class="icon-link footer__chain" />
-            </template>
-          </base-btn>
-          <base-field
-            v-model="subCommentInput"
-            class="footer__input"
-            :placeholder="$t('discussions.input')"
-            rules="required|max:250"
-            :name="$t('discussions.response')"
-            mode="comment-field"
-            @keyup.enter.native="handleSubmit(addSubCommentResponse(data, 5))"
-          />
-          <base-btn
-            class="footer__btn"
-            :disabled="!validated || !passed || invalid"
-            @click="handleSubmit(addSubCommentResponse(data, 5))"
-          >
-            <template v-slot:left>
-              <span class="icon-send footer__arrow" />
-            </template>
-          </base-btn>
-        </div>
-      </validation-observer>
+      <comment-footer
+        v-if="isReply"
+        :comment="data"
+        :level="5"
+      />
     </div>
     <div
       v-for="(sub5) in filterComments(sub5Comments, data.id)"
@@ -274,13 +243,6 @@ export default {
     width: 100%;
     display: flex;
   }
-  &__footer {
-    height: 40px;
-    display: grid;
-    grid-template-columns: 1fr 11fr 1fr;
-    margin: 20px;
-    align-items: center;
-  }
   &_sub4 {
     //background: #505052;
     margin-left: 45px;
@@ -364,54 +326,6 @@ export default {
     &_right {
       margin: 7px 7px 7px 0;
     }
-  }
-}
-.footer {
-  animation: show  1s 1;
-  display: flex;
-  &__input {
-    @include text-usual;
-    width: 100%;
-    height: 40px;
-    border-radius: 6px;
-    border: none;
-    padding: 10px 15px 10px 15px;
-    margin: 0 0 20px 0;
-  }
-  &__chain {
-    padding: 0 0 0 5px;
-    display: flex;
-    width: 40px;
-    height: 40px;
-    background: #F7F8FA;
-    border-radius: 6px;
-    align-items: center;
-    justify-content: center;
-    color: #000000;
-    font-size: 25px;
-    cursor: pointer;
-  }
-  &__btn {
-    width: 40px;
-    height: 40px;
-    background: #F7F8FA;
-    cursor: pointer;
-    &:hover {
-      background: #F7F8FA;
-    }
-  }
-  &__arrow {
-    display: flex;
-    width: 40px;
-    height: 40px;
-    background: #F7F8FA;
-    border-radius: 6px;
-    align-items: center;
-    justify-content: center;
-    font-size: 25px;
-    color: #0083C7;
-    cursor: pointer;
-    padding: 0 0 0 10px;
   }
 }
 </style>
