+++ components/app/Pages/Discussions/CommentField4.vue
@@ -182,13 +182,6 @@ export default {
     width: 100%;
     display: flex;
   }
-  &__footer {
-    height: 40px;
-    display: grid;
-    grid-template-columns: 1fr 11fr 1fr;
-    margin: 20px;
-    align-items: center;
-  }
   &_sub5 {
     //background: #37373a;
     margin-left: 60px;
@@ -258,54 +251,6 @@ export default {
     &_right {
       margin: 7px 7px 7px 0;
     }
-  }
-}
-.footer {
-  animation: show  1s 1;
-  display: flex;
-  &__input {
-    @include text-usual;
-    width: 100%;
-    height: 40px;
-    border-radius: 6px;
-    border: none;
-    padding: 10px 15px 10px 15px;
-    margin: 0 0 20px 0;
-  }
-  &__chain {
-    padding: 0 0 0 5px;
-    display: flex;
-    width: 40px;
-    height: 40px;
-    background: #F7F8FA;
-    border-radius: 6px;
-    align-items: center;
-    justify-content: center;
-    color: #000000;
-    font-size: 25px;
-    cursor: pointer;
-  }
-  &__btn {
-    width: 40px;
-    height: 40px;
-    background: #F7F8FA;
-    cursor: pointer;
-    &:hover {
-      background: #F7F8FA;
-    }
-  }
-  &__arrow {
-    display: flex;
-    width: 40px;
-    height: 40px;
-    background: #F7F8FA;
-    border-radius: 6px;
-    align-items: center;
-    justify-content: center;
-    font-size: 25px;
-    color: #0083C7;
-    cursor: pointer;
-    padding: 0 0 0 10px;
   }
 }
 </style>
