+++ components/app/Pages/Discussions/RootCommentField.vue
@@ -71,41 +71,10 @@
         />
       </div>
     </div>
-    <validation-observer
-      v-slot="{handleSubmit, validated, passed, invalid}"
-    >
-      <!--  Bottom -->
-      <div
-        class="comment__footer footer"
-      >
-        <base-btn
-          class="footer__btn hide"
-          :disabled="!validated || !passed || invalid"
-        >
-          <template v-slot:left>
-            <span class="icon-link footer__chain" />
-          </template>
-        </base-btn>
-        <base-field
-          v-model="subCommentInput"
-          class="footer__input"
-          :placeholder="$t('discussions.input')"
-          rules="required|max:250"
-          :name="$t('discussions.response')"
-          mode="comment-field"
-          @keyup.enter.native="handleSubmit(addSubCommentResponse(comment, 2))"
-        />
-        <base-btn
-          class="footer__btn"
-          :disabled="!validated || !passed || invalid"
-          @click="handleSubmit(addSubCommentResponse(comment, 2))"
-        >
-          <template v-slot:left>
-            <span class="icon-send footer__arrow" />
-          </template>
-        </base-btn>
-      </div>
-    </validation-observer>
+    <comment-footer
+      :comment="comment"
+      :level="2"
+    />
   </div>
 </template>
 
@@ -125,7 +94,6 @@ export default {
       subCommentsOnPage: 5,
       comments: [],
       sub2Comments: [],
-      subCommentInput: '',
       count: 1,
     };
   },
@@ -137,28 +105,6 @@ export default {
   methods: {
     clearSubs() {
       this.sub2Comments = [];
-    },
-    async addSubCommentResponse(comment, level) {
-      if (!comment.rootCommentId) {
-        const payload = {
-          rootCommentId: comment.id,
-          text: this.subCommentInput,
-          medias: [],
-        };
-        await this.$store.dispatch('discussions/sendCommentOnDiscussion', { id: this.currentDiscussion.id, payload });
-        await this.loadSubs(comment.id, level);
-        this.subCommentInput = '';
-      } else {
-        const payload = {
-          rootCommentId: comment.id,
-          text: this.subCommentInput,
-          medias: [],
-        };
-        await this.$store.dispatch('discussions/sendCommentOnDiscussion', { id: this.currentDiscussion.id, payload });
-        await this.loadSubs(comment.rootCommentId, level);
-        this.isReply = false;
-        this.subCommentInput = '';
-      }
     },
     async toggleLikeOnComment(comment) {
       await this.$store.dispatch('discussions/toggleLikeOnComment', { id: comment.id, like: comment && !Object.keys(comment.commentLikes).length > 0 });
@@ -243,13 +189,6 @@ export default {
     word-break: break-all;
     width: 100%;
     display: flex;
-  }
-  &__footer {
-    height: 40px;
-    display: grid;
-    grid-template-columns: 1fr 11fr 1fr;
-    margin: 20px;
-    align-items: center;
   }
   &_sub2 {
     background: #8D96A2;
@@ -349,53 +288,6 @@ export default {
         color: #E9EDF2;
       }
     }
-  }
-}
-.footer {
-  animation: show  1s 1;
-  display: flex;
-  &__input {
-    @include text-usual;
-    width: 100%;
-    height: 40px;
-    border-radius: 6px;
-    border: none;
-    padding: 10px 15px 10px 15px;
-    margin: 0 0 20px 0;
-  }
-  &__chain {
-    padding: 0 0 0 5px;
-    display: flex;
-    width: 40px;
-    height: 40px;
-    background: #F7F8FA;
-    border-radius: 6px;
-    align-items: center;
-    justify-content: center;
-    color: #000000;
-    font-size: 25px;
-    cursor: pointer;
-  }
-  &__btn {
-    width: 40px;
-    height: 40px;
-    background: #F7F8FA;
-    cursor: pointer;
-    &:hover {
-      background: #F7F8FA;
-    }
-  }
-  &__arrow {
-    display: flex;
-    width: 40px;
-    height: 40px;
-    border-radius: 6px;
-    align-items: center;
-    justify-content: center;
-    font-size: 25px;
-    color: #0083C7;
-    cursor: pointer;
-    padding: 0 0 0 10px;
   }
 }
 </style>
